+++ content/lessons/chapter-1/genesis-3.tsx
@@ -1,49 +1,32 @@
 'use client'
 
 import { useLocalizedRoutes, useTranslations } from 'hooks'
-import { CodeExample, LessonInfo, Title, Text, TerminalChallenge } from 'ui'
+import { Button, ChapterIntro } from 'ui'
 
-export default function Genesis5({ lang }) {
+export default function Genesis2({ lang }) {
   const routes = useLocalizedRoutes()
   const t = useTranslations(lang)
 
   return (
-    <TerminalChallenge
-      expectedInput={{
-        userVariable: '$scriptSigHex',
-        value:
-          '04ffff001d0104455468652054696d65732030332f4a616e2f32303039204368616e63656c6c6f72206f6e206272696e6b206f66207365636f6e64206261696c6f757420666f722062616e6b73',
-      }}
-      saveInfo={{
-        chapter: 'chapter-1',
-        challenge: 'transacting-1',
-      }}
-      next={`${routes.chaptersUrl}/chapter-1/genesis-4`}
-      successMessage={t('chapter_one.genesis_three.success')}
-      customLines={t('chapter_one.genesis_three.terminal_challenge_lines')}
-    >
-      <LessonInfo>
-        <Title>{t('chapter_one.genesis_three.heading')}</Title>
+    <ChapterIntro heading={t('chapter_one.genesis_one.heading')}>
+      <p className="mt-2 text-lg md:text-xl">
+        {t('chapter_one.genesis_one.paragraph_one')}
+      </p>
 
-        <Text className="mt-2 text-lg md:text-xl">
-          {t('chapter_one.genesis_three.paragraph_one')}
-        </Text>
+      <p className="mt-8 text-lg md:text-xl">
+        {t('chapter_one.genesis_one.paragraph_two')}
+      </p>
 
-        <Text className="mt-2 text-lg md:text-xl">
-          {t('chapter_one.genesis_three.paragraph_two')}
-        </Text>
-
-        <CodeExample
-          className="mt-4"
-          code={`echo $scriptSigHex | xxd -r -p`}
-          language="shell"
-          copy
-        />
-      </LessonInfo>
-    </TerminalChallenge>
+      <Button
+        href={`${routes.chaptersUrl}/chapter-1/genesis-4`}
+        classes="mt-10 max-md:w-full"
+      >
+        {t('shared.start')}
+      </Button>
+    </ChapterIntro>
   )
 }
 
 export const metadata = {
-  title: 'chapter_one.genesis_three.title',
+  title: 'chapter_one.genesis_one.title',
 }
